move from double to long in order to get rid of error of calculation

+++ services/recommender_service/recommender-real-time-service/src/main/java/vn/datnguyen/recommender/Repository/BoundedRingInterface.java
@@ -8,14 +8,14 @@ public interface BoundedRingInterface {
     public final static String BOUNDED_RING_ROW = "bounded_ring_row";
     public final static String RING_ID = "ring_id"; // UUID
     public final static String CENTRE_ID = "centre_id"; //int
-    public final static String LOWER_BOUND_RANGE = "lower_bound_range"; //double
-    public final static String UPPER_BOUND_RANGE = "upper_bound_range"; // double
+    public final static String LOWER_BOUND_RANGE = "lower_bound_range"; //long
+    public final static String UPPER_BOUND_RANGE = "upper_bound_range"; // long
     public final static String CAPACITY = "capacity"; // int
 
     SimpleStatement createRowIfNotExists();
-    SimpleStatement addNewBoundedRing(UUID ringId, int centreId, double lbRange, double ubRange, int capacity);
+    SimpleStatement addNewBoundedRing(UUID ringId, int centreId, long lbRange, long ubRange, int capacity);
     SimpleStatement findBoundedRingById(UUID ringId, int centreId);
-    SimpleStatement findBoundedRingByIdAndRange(int centreId, double ubRange);
+    SimpleStatement findBoundedRingByIdAndRange(int centreId, long ubRange);
     SimpleStatement findAllBoundedRingInCentre(int centreId);
     //SimpleStatement updateBoundedRingCapacityById(UUID ringId, int centreId, double ubRange, int capacity);
     SimpleStatement deleteBoundedRingById(UUID ringId, int centreId);
